src/requirements.txt
# Originial Dependencies

#-e git+https://github.com/GeoNode/geonode-mapstore-client.git@master#egg=django_geonode_mapstore_client
#-e git+https://github.com/GeoNode/geonode-importer.git@master#egg=geonode-importer
#-e git+https://github.com/GeoNode/geonode.git@master#egg=GeoNode

# Django 4.2 Dependencies

git+https://github.com/GeoNode/dynamic-rest.git@master#egg=dynamic_rest
git+https://github.com/GeoNode/pinax-notifications.git@django_upgrade#egg=pinax-notifications
git+https://github.com/GeoNode/geonode-mapstore-client.git@django_upgrade#egg=django_geonode_mapstore_client
git+https://github.com/GeoNode/geonode-importer.git@django_upgrade#egg=geonode-importer
git+https://github.com/GeoNode/geonode-user-messages.git@django_upgrade#egg=geonode-user-messages
git+https://github.com/GeoNode/geonode-announcements.git@django_upgrade#egg=geonode-announcements

-e git+https://github.com/roicort/geonode.git@master#egg=GeoNode

# Wagtail 5 Dependencies

beautifulsoup4==4.11.2
django-taggit==4.0.0
Wagtail>=5.0

